+++ src/lib/index.ts
@@ -1,5 +1,5 @@
 // Reexport your entry components here
 export type {WritableStorage} from "$lib/store/shared/WritableStorage";
-export {writableSessionStorage} from "$lib/store/sessionstorage/WritableSessionStorage";
-export {writableLocalStorage} from "$lib/store/localstorage/WritableLocalStorage";
+export {writableSessionStorage} from "$lib/store/session/WritableSessionStorage";
+export {writableLocalStorage} from "$lib/store/local/WritableLocalStorage";
 export {writableAsync, type WritableAsync} from "$lib/store/async/WritableAsync";
