+++ src/lib/sandbox/stores/historic.ts
@@ -1,4 +1,4 @@
-import {historicWritable} from "$lib/store/shared/HistoricWritable";
+import {historicWritable} from "$lib/store/historic/HistoricWritable";
 
 /** @internal */
 export const historic = historicWritable<string>("initial value");
